requirements.txt
langchain
langgraph
openai
python-dotenv
requests
beautifulsoup4
